Integrate Google ReCaptcha
Added unique props ID for each instance of <ContactForm /> on the 'More Info' page so that the reCaptchaIsValid state is tied to that instance of the form for proper form validation

Also checked for height of the reCaptcha <div>, because after successful validation, Google's script hides the element from the page for the remainder of the session. By checking the height, I know to set the reCaptchaIsValid boolean to true if the reCaptcha is not present and adjust margins.