Fix live 'E-mail Sign-Up' page; make 'Thank You' page standalone
Updated file structure and links in e-mail sign-up page files to reflect file structure on portfolio website.

Recaptcha wasn't displaying on the 'More Info' page after the user submitted the form once. This is because my website is really just one page (SPA) with React. Making the Thank You page, which the user is forwarded to on form submit, a separate page forces Google to reset its reCaptcha on the form.

Updated Resume PDF.